[EDELIVERY-6521] avoid md2-datepicker/mat-select conflicts
md2-datepicker creates a 'popupRef' object (represented in the DOM by a <div> with class="cdk-overlay-container") when opening for the first time, hides it when closing, and reuses it when opening again.

OTOH, mat-select will reuse any existing cdk-overlay-container when opening, even if it was created by some other component. This means it may alter the cdk-overlay-container instance created by md2-datepicker.

If a "cdk-overlay-container" div is already present on the page, mat-select will pick it and use it and leave the md2-datepicker instances alone.
 