EDELIVERY-8602 Migration of Primary Keys From the v4.2.x Format to the New v5.0 Format
Only use the data scripts in scratch mode to avoid collision with existing migration data.
Fix waitForDatabase to look for the rightmost digit when comparing the ID_PK.
Force the nullability off for the REV_ID and ID columns of the V_AUDIT_DETAIL to prevent random failures in the db migration (anyway these column values can't be null).